+++ src/main/java/net/kyrptonaught/linkedstorage/block/StorageBlockRenderer.java
@@ -32,7 +32,7 @@ public class StorageBlockRenderer extends BlockEntityRenderer<StorageBlockEntity
 
     @Override
     public void render(StorageBlockEntity blockEntity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
-        int[] dyes = blockEntity.getChannel();
+        byte[] dyes = blockEntity.getChannel();
         float[] color1 = DyeColor.byId(dyes[0]).getColorComponents();
         float[] color2 = DyeColor.byId(dyes[1]).getColorComponents();
         float[] color3 = DyeColor.byId(dyes[2]).getColorComponents();
